backend/requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
sse-starlette>=2.1.0

# LangGraph + LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
langchain-google-genai>=2.0.0
langchain-mcp-adapters>=0.1.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.0

# Redis + Celery
redis>=5.0.0
celery[redis]>=5.4.0
flower>=2.0.0

# Auth / Config
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0

# Tools — Search & Web
tavily-python>=0.3.0
exa-py>=1.1.0
firecrawl-py>=1.4.0
playwright>=1.44.0
google-search-results>=2.4.2

# Tools — Social & Community
praw>=7.7.0
google-api-python-client>=2.130.0

# Tools — Trends & Market
pytrends>=4.9.2
newsapi-python>=0.2.7

# Tools — Temporal / Calendar
requests>=2.32.0
httpx>=0.27.0
aiohttp>=3.9.0

# Utilities
python-dateutil>=2.9.0
tenacity>=8.3.0
structlog>=24.1.0
